Refactor intake command from CommandGroup to work directly with subsystems. Fixes #29
