+++ src/main/java/org/frc571/bradley/commands/IntakeCommand.java
@@ -1,18 +1,54 @@
 
 package org.frc571.bradley.commands;
 
+import edu.wpi.first.wpilibj2.command.CommandBase;
+
 import org.frc571.bradley.subsystems.Index;
 import org.frc571.bradley.subsystems.Intake;
+import org.frc571.bradley.Constants.SpeedConstants;
 
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
+public class IntakeCommand extends CommandBase {
 
-public class IntakeCommand extends SequentialCommandGroup {
+    private final Index index;
+    private final Intake intake;
 
     public IntakeCommand() {
 
-        addCommands(
-                new SuckCommand(),
-                new IndexCommand());
+        index = Index.getInstance();
+        addRequirements(index);
+        intake = Intake.getInstance();
+        addRequirements(index);
+
+    }
+
+    // Called when the command is initially scheduled.
+    @Override
+    public void initialize() {
+        intake.setSpeed(SpeedConstants.kIntakeMotorSpeed);
+    }
+
+    // Called every time the scheduler runs while the command is scheduled.
+    @Override
+    public void execute() {
+        if (index.getFrontInput()) {
+            index.turn();
+        }
+    }
+
+    // Called once the command ends or is interrupted.
+    @Override
+    public void end(boolean interrupted) {
+        index.stop();
+        intake.stop();
+    }
+
+    // Returns true when the command should end.
+    @Override
+    public boolean isFinished() {
+        if (index.getMiddleInput() || index.getBackInput()) {
+            return true;
+        }
+        return false;
     }
 
     @Override
